// apps/CementAgency/src/app/models/accounts.model.ts
    export class Accounts {
      AccountID: string = '';
      AcctCat: string = '';
      AccountChart: string = '';
      AccountCode?: any;
      AccountName: string = '';
      ContactPerson: string = '';
      Address: string = '';
      PhoneNo: string = '';
      City: string = '';
      OBalance: number = 0 ;
      Balance: number = 0;
      Remarks: string = '';
      Commission: number = 0;
      MobileNo: string = '';
      WhatsappNo: string = '';
      Email: string = '';
      CreditLimit: string = '';
      Folio: string = '';
      SalesmanID?: any;
      Status: string = '';
      NTNNo: string = '';
  }